backend/requirements.txt
fastapi
uvicorn
pymongo
pandas
pdfplumber
scikit-learn
python-dotenv
ollama
nomic
jobspy
langchain
requests
python-dotenv